requirements.txt
arrow==1.3.0
asgiref==3.8.1
babel==2.16.0
binaryornot==0.4.4
boto3==1.34.23
botocore==1.34.162
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cookiecutter==2.6.0
crispy-bootstrap5==2024.2
crispy-tailwind==1.0.3
cssselect2==0.7.0
diff-match-patch==20241021
Django==5.0.7
django-admin-rangefilter==0.13.2
django-browser-reload==1.17.0
django-ckeditor==6.7.2
django-crispy-forms==2.1
django-csp==4.0
django-daterange-filter==1.3.0
django-daterangefilter==1.0.1
django-filter==24.2
-e git+https://github.com/django-import-export/django-import-export.git@b3cff82a23df5815204661780a85e74126fb63ea#egg=django_import_export
django-ipware==7.0.1
django-js-asset==3.0.1
django-social-share==2.3.0
django-storages==1.14.4
django-tailwind==3.8.0
docxcompose==1.4.0
docxtpl==0.19.1
fonttools==4.55.3
humanize==4.10.0
idna==3.10
Jinja2==3.1.5
jmespath==1.0.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysqlclient==2.2.6
packaging==25.0
pillow==11.1.0
psycopg2-binary==2.9.9
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.1
pyphen==0.17.2
python-dateutil==2.8.2
python-decouple==3.8
python-docx==1.1.2
python-ipware==3.0.0
python-slugify==8.0.4
pytz==2023.3.post1
PyYAML==6.0.2
reportlab==4.2.5
requests==2.32.3
rich==13.9.4
s3transfer==0.10.0
setuptools==75.8.0
six==1.16.0
sqlparse==0.5.0
tablib==3.8.0
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
weasyprint==63.1
webencodings==0.5.1
wfastcgi==3.0.0
zopfli==0.2.3.post1
